feat(VT-50): adiciona semifinal dto e coloca dtos em tabela.dto
